mc6ul_m30_defconfig: enable "run" command
fixes DWI-25:https://mentorgraphics.atlassian.net/browse/DWI-25

This is needed to overcome a bug in the u-boot when trying to use
directly the bootm command(bootm ${loadaddr}:kernel@1 - ${loadaddr}:fdt@1)
By using the bootm directly the kernel is resetting. Did not find the cause
till now for this behavior.

Signed-off-by: Nicusor Huhulea <nicusor_huhulea@mentor.com>